third_party/blink/web_tests/fast/canvas/canvas-ImageBitmap-transferable-expected.txt

Tests that ImageBitmap is transferable and that the pixel data survives the trip, as well as createImageBitmap(ImageBitmap) works on worker thread

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS bitmapWidth is imageWidth
PASS bitmapHeight is imageHeight
PASS bitmapWidth is 0
PASS bitmapHeight is 0
PASS Apply structured cloning to a neutered ImageBitmap should throw an exception: DataCloneError: Failed to execute 'postMessage' on 'Worker': An ImageBitmap is detached and could not be cloned.
PASS createImageBitmap from a neutered ImageBitmap was rejected
PASS bitmapWidth is imageWidth
PASS bitmapHeight is imageHeight
PASS newImage.data[0] is within 5 of image.data[0]
PASS newImage.data[1] is within 5 of image.data[1]
PASS newImage.data[2] is within 5 of image.data[2]
PASS newImage.data[3] is image.data[3]
PASS newImage.data[4] is within 5 of image.data[4]
PASS newImage.data[5] is within 5 of image.data[5]
PASS newImage.data[6] is within 5 of image.data[6]
PASS newImage.data[7] is image.data[7]
PASS newImage.data[8] is within 5 of image.data[8]
PASS newImage.data[9] is within 5 of image.data[9]
PASS newImage.data[10] is within 5 of image.data[10]
PASS newImage.data[11] is image.data[11]
PASS newImage.data[12] is within 5 of image.data[12]
PASS newImage.data[13] is within 5 of image.data[13]
PASS newImage.data[14] is within 5 of image.data[14]
PASS newImage.data[15] is image.data[15]
PASS newImage.data[16] is within 5 of image.data[16]
PASS newImage.data[17] is within 5 of image.data[17]
PASS newImage.data[18] is within 5 of image.data[18]
PASS newImage.data[19] is image.data[19]
PASS newImage.data[20] is within 5 of image.data[20]
PASS newImage.data[21] is within 5 of image.data[21]
PASS newImage.data[22] is within 5 of image.data[22]
PASS newImage.data[23] is image.data[23]
PASS successfullyParsed is true

TEST COMPLETE
